requirements.txt
requests==2.18.4
vcrpy==1.12.0
pytest==3.6.1
